requirements.txt
BeautifulSoup==3.2.1
Django==1.7.3
Fabric==1.9.1
MySQL-python==1.2.5
PyYAML==3.11
<<<<<<< HEAD
SQLAlchemy==0.9.7
=======
SQLAlchemy==0.9.8
>>>>>>> master
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
boto==2.28.0
celery==3.1.17
certifi==14.05.14
csvkit==0.9.0
dbf==0.94.3
django-admin-bootstrapped==2.1.1
django-bakery==0.7.1
django-debug-toolbar==1.2.2
django-mass-edit==2.2
django-tastypie==0.12.0
ecdsa==0.11
enum34==1.0
hurry.filesize==0.9
jdcal==1.0
kombu==3.0.24
<<<<<<< HEAD
openpyxl==2.1.0
=======
latimes-calculate==0.3.0
openpyxl==2.1.4
>>>>>>> master
paramiko==1.14.1
progressbar==2.2
pycrypto==2.6.1
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.7
requests==2.2.1
<<<<<<< HEAD
six==1.5.2
=======
six==1.9.0
>>>>>>> master
sqlparse==0.1.14
twitter==1.15.0
webassets==0.10.1
wsgiref==0.1.2
xlrd==0.9.3
yolk==0.4.3
